reorganizing project directories, adding unit test
